Fix build warnings that get treated as error

+++ cpp/src/arrow/flight/sql/odbc/flight_sql/address_info.cc
@@ -34,7 +34,7 @@ bool AddressInfo::GetAddressInfo(const std::string& host, char* host_name_info,
   }
 
   error = getnameinfo(addrinfo_result_->ai_addr, addrinfo_result_->ai_addrlen,
-                      host_name_info, max_host, NULL, 0, 0);
+                      host_name_info, static_cast<DWORD>(max_host), NULL, 0, 0);
   return error == 0;
 }
